# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Gia Phả Dòng Họ API
  version: 1.0.0-sprint1
  summary: REST API cho hệ thống gia phả dòng họ (Giai đoạn 1 / Phase 1 MVP).
  description: |-
    ## Tiếng Việt

    API REST của hệ thống quản lý gia phả. Truy vấn cây **lồng sâu** dùng GraphQL
    (`POST /api/v1/graphql`, schema tại `contracts/schema.graphqls`); mọi thứ còn lại dùng REST.

    Quy ước chung:

    * **Versioning theo URL**: mọi đường dẫn nằm dưới `/api/v1`.
    * **Lỗi theo RFC 7807** Problem Details, `Content-Type: application/problem+json`.
      Trường mở rộng `code` là mã máy đọc được — FE phân nhánh theo `code`, **không** theo `detail`.
    * **Phân trang thống nhất** cho mọi endpoint trả danh sách: query `page` (0-based) + `size`,
      phản hồi bọc trong `{ items: [...], page: { ... } }`.
    * **Định danh**: UUID v4/v7 dạng chuỗi.
    * **Thời điểm**: ISO-8601 có offset (`2026-03-14T08:00:00+07:00`). Múi giờ nghiệp vụ là **GMT+7**.
    * **Ngày song lịch**: mọi mốc sinh/mất đều có cả dương lịch (`solar`) và âm lịch (`lunar`).
      `death.lunar` là **nguồn chân lý** để tính ngày giỗ.
    * **Ngôn ngữ**: gửi `Accept-Language: vi` hoặc `en`; mặc định `vi`.
      Danh xưng (`title`) luôn trả bằng tiếng Việt vì đó là dữ liệu nghiệp vụ, không phải chuỗi i18n.

    ### Phân tầng riêng tư — đọc kỹ trước khi code FE (Nghị định 13/2023, BA v2 §10)

    Cùng một `PersonDto` nhưng **tập trường trả về khác nhau** theo vai người gọi, `isAlive`
    và `privacyLevel` của chính người đó:

...

...
servers:
...
security:
  - bearerAuth: []
  - {}
paths:
...
  /push/public-key:
    get:
      tags: [push]
      operationId: getVapidPublicKey
      summary: Khóa công khai VAPID / VAPID public key
      description: |-
        Trả **khóa công khai VAPID** dạng base64url không padding, để service worker truyền vào
        `PushManager.subscribe({ applicationServerKey })`.

        Không có endpoint này thì FE buộc phải nhúng cứng khóa vào build — đổi khóa là phải build lại,
        và môi trường dev/staging/prod dùng khóa khác nhau sẽ lệch. **Khóa bí mật VAPID không bao giờ
        rời server** và không xuất hiện ở bất kỳ phản hồi nào.
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VapidPublicKey'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /push/subscriptions:
    post:
      tags: [push]
      operationId: registerPushSubscription
      summary: Đăng ký thiết bị nhận Web Push / Register a Web Push subscription
      description: |-
        Đăng ký một `PushSubscription` chuẩn **VAPID / Web Push (RFC 8291)** cho người đang đăng nhập.
        Body chính là kết quả `PushSubscription.toJSON()` của trình duyệt, cộng vài trường phụ trợ.

        **Idempotent theo `endpoint`**: gửi lại cùng `endpoint` sẽ **cập nhật** bản ghi cũ và trả `200`
        (không phải `201`, không tạo bản trùng). Mỗi người có thể có nhiều thiết bị.

        **Ghi chú nền tảng (FE phải xử lý, đừng để im lặng thất bại):**

        * iOS chỉ cho Web Push khi PWA **đã được cài vào màn hình chính** (iOS 16.4+). App cần
          hướng dẫn "Thêm vào màn hình chính", nếu không phần lớn người dùng iPhone sẽ không bao giờ nhận push.
        * Xin quyền **đúng lúc** (sau khi người dùng xem một hồ sơ hoặc một ngày giỗ), không phải ngay khi vào trang.
        * Người từ chối quyền vẫn nhận đủ thông báo in-app — push không phải điều kiện tiên quyết.

        Khi push gateway trả `404`/`410` (subscription bị thu hồi hoặc hết hạn), backend **tự xóa**
        bản ghi và không retry. FE có thể thấy một subscription biến mất mà không do mình gọi `DELETE`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSubscriptionCreateRequest'
      responses:
        '201':
          description: Đã đăng ký thiết bị mới. / New subscription registered.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSubscriptionDto'
        '200':
          description: Đã cập nhật đăng ký sẵn có cùng `endpoint`. / Existing subscription updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSubscriptionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /push/subscriptions/{id}:
    delete:
      tags: [push]
      operationId: deletePushSubscription
      summary: Hủy đăng ký Web Push / Remove a Web Push subscription
      description: |-
        Xóa một đăng ký push của **chính mình** (người dùng tắt công tắc trong phần cài đặt,
        hoặc gọi `pushSubscription.unsubscribe()`).

        Đây là **xóa cứng** — quy tắc "chỉ xóa mềm" áp cho nhân khẩu trong gia phả, không áp cho
        dữ liệu kỹ thuật của thiết bị.

        Đăng ký không tồn tại hoặc thuộc người khác ⇒ `404`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Id bản ghi đăng ký (không phải `endpoint`). / Subscription record id.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Đã hủy đăng ký. / Subscription removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # VẬN HÀNH NHẮC GIỖ / REMINDER PIPELINE OPERATIONS (System Admin)
  # ---------------------------------------------------------------------------
  /admin/reminders/generate:
    post:
      tags: [admin-reminders]
      operationId: generateReminders
      summary: Sinh lịch nhắc giỗ ngay / Generate reminder jobs now
      description: |-
        Chạy ngay job sinh `reminder_job` thay vì chờ scheduler 01:30 (GMT+7).

        **Vì sao có endpoint này.** Trên môi trường demo/staging vốn được dựng lên rồi tắt trong
        ngày, bảng `reminder_job` luôn rỗng và cả đường ống nhắc giỗ không trình diễn được
        đầu-cuối. Bấm nút vẫn tốt hơn sửa cron thành mỗi phút rồi quên trả lại.

        **Idempotent.** Gọi lại với cùng `referenceDate` trả `200` và `createdJobs = 0`; chống
        trùng do chỉ mục duy nhất `ux_reminder_job_occurrence (event_id, occurrence_year,
        offset_days)` bảo đảm — không phải do một phép kiểm tra ở tầng ứng dụng. `scannedEvents`
        vẫn giữ nguyên, nhờ đó phân biệt được "job đã có sẵn từ lượt trước" với "không có sự kiện
        nào để quét".

        **`referenceDate` đổi luôn cả mốc "bây giờ"** thành 00:00 của chính ngày đó. Nếu giữ thời
        điểm thật thì mọi mốc nhắc nằm trước hiện tại sẽ bị loại, và tham số coi như vô dụng đúng ở
        trường hợp cần nó nhất: dựng lại cả ba mốc **7 / 3 / 1** để xem chúng chạy thật.

        **Hệ quả phải biết trước:** `referenceDate` trong quá khứ sinh ra job có `fire_at` đã qua,
        và lượt đẩy kế tiếp (5 phút một lần, hoặc `POST /admin/reminders/dispatch`) sẽ **gửi thật**
        tới người trong họ. Đó là lý do endpoint này chỉ dành cho System Admin.

        Chỉ vai **System Admin** (`ADMIN`) — vai kỹ thuật toàn hệ thống, không phải chức danh dòng
        tộc. Cố ý không mở cho Hội đồng Tộc biểu: hội đồng là thẩm quyền *nội dung* của dòng họ,
        đây là bề mặt *vận hành*. Cũng không có phạm vi chi/ngành: job quét toàn bộ dòng họ nên
        không có `ltree` nào để giới hạn nó.
      security:
        - bearerAuth: []
      parameters:
        - name: referenceDate
          in: query
          description: |-
            Ngày được coi là "hôm nay" (GMT+7). Mặc định là hôm nay.
            Chỉ chấp nhận lệch tối đa **366 ngày** về mỗi phía: gõ nhầm năm mà không có chốt này
            thì job sẽ quét toàn bộ sự kiện của dòng họ để sinh lịch nhắc cho một mốc chẳng ai dùng.
          schema:
            type: string
            format: date
          example: '2026-09-05'
      responses:
        '200':
          description: Đã chạy xong (kể cả khi không tạo job nào mới).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderGenerationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: |-
            Đã có một lượt sinh khác đang chạy trên instance này
            (`code` = `events.reminder-generation-in-progress`). Không phải lỗi dữ liệu — đợi lượt
            trước xong rồi bấm lại. Chốt này chỉ để một lần bấm sốt ruột không nhân đôi tải quét
            toàn bảng; việc ghi vốn đã idempotent.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/reminders/dispatch:
    post:
      tags: [admin-reminders]
      operationId: dispatchDueReminders
      summary: Đẩy lịch nhắc đã tới giờ lên hàng đợi / Dispatch due reminders now
      description: |-
        Đẩy ngay các `reminder_job` đã tới giờ lên RabbitMQ, thay vì chờ lượt quét 5 phút.

        **Luồng HTTP không chờ gửi.** Ở đây chỉ ghi tin vào broker rồi trả về; việc gọi gateway
        (in-app, Web Push) diễn ra ở consumer. Gateway chết cũng không giữ lại request nào. Vì vậy
        `dispatchedJobs` là **số job đã đẩy đi**, không phải số người đã nhận được.

        Job có ngày giỗ **đã qua** không được gửi mà chuyển `CANCELLED` — nhắc một cái giỗ đã cúng
        xong là vô nghĩa và gây hoang mang. Job có sự kiện đã xóa mềm cũng vậy.

        Chỉ vai **System Admin** (`ADMIN`).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDispatchResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # GRAPHQL
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        JWT do **Keycloak** (OAuth2/OIDC) cấp, gồm cả social login Google/Zalo.
        Backend map `keycloak_sub → app_user → person`.

        Vai trò (`realm_access.roles`): `ADMIN` (kỹ thuật, toàn hệ thống) ·
        `COUNCIL` (Hội đồng Tộc biểu / Tộc trưởng, toàn dòng họ) ·
        `BRANCH_HEAD` (Trưởng Chi/Ngành — **chỉ trong chi được giao**) ·
        `MEMBER` · không token = Khách (`GUEST`).

        **Kiểm quyền hai chiều**: vai trò **và** phạm vi chi/ngành theo `ltree` của đối tượng.
        Có vai `BRANCH_HEAD` không đồng nghĩa được sửa mọi người trong họ.

        Chức danh dòng tộc (Tộc trưởng, Trưởng chi — theo huyết thống/đích tôn) là dữ liệu nghiệp vụ,
        **tách khỏi** vai kỹ thuật trong token; một người có thể giữ cả hai.
...
  responses:
    BadRequest:
      description: Tham số hoặc body không hợp lệ. / Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    Unauthorized:
      description: Thiếu hoặc sai token. / Missing or invalid token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |-
        Đã xác thực nhưng không đủ quyền — sai vai trò **hoặc** đối tượng nằm ngoài phạm vi
        chi/ngành được giao (`BRANCH_SCOPE_VIOLATION`).

        Lưu ý: với **người còn sống bị ẩn theo phân tầng**, API trả `404` chứ không `403`,
        để không xác nhận rằng người đó tồn tại.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: |-
        Không tồn tại, **hoặc** người gọi không được phép biết là nó tồn tại
        (người còn sống với Khách, nhân khẩu đã xóa mềm với người không phải quản trị,
        thông báo của người khác). Hai trường hợp cố ý không phân biệt được.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: Body đúng cú pháp nhưng vi phạm ràng buộc nghiệp vụ. / Business rule violation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    TooManyRequests:
      description: Vượt hạn mức gọi. / Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Số giây nên chờ trước khi thử lại. / Seconds to wait.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: |-
        Lỗi hệ thống. `detail` **không bao giờ** chứa stack trace hay giá trị dữ liệu;
        dùng `traceId` để tra log (Loki/OTel).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      title: Problem (RFC 7807)
      description: |-
        Định dạng lỗi dùng chung cho **toàn bộ** API, theo RFC 7807 Problem Details,
        trả với `Content-Type: application/problem+json`.

        FE phân nhánh xử lý theo `code` (mã máy đọc, ổn định), **không** theo `title`/`detail`
        (là văn bản cho người, đổi theo `Accept-Language`).
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: URI định danh loại lỗi. / URI identifying the problem type.
          default: about:blank
        title:
          type: string
          description: Tiêu đề ngắn cho người đọc. / Short human-readable summary.
        status:
          type: integer
          description: Mã HTTP, lặp lại để tiện log. / HTTP status code.
        detail:
          type: string
          description: |-
            Diễn giải cho người dùng cuối. **Không bao giờ** chứa dữ liệu nhạy cảm (SĐT, email,
            địa chỉ), stack trace, hay câu SQL/Cypher.
        instance:
          type: string
          format: uri-reference
          description: Đường dẫn đã gọi. / The request path.
        code:
          $ref: '#/components/schemas/ProblemCode'
        traceId:
          type: string
          description: |-
            Trace id (W3C traceparent) để đối chiếu log OTel/Loki. FE nên hiện cho người dùng khi
            báo lỗi hệ thống. / Trace id for log correlation.
        timestamp:
          type: string
          format: date-time
          description: Thời điểm phát sinh lỗi (GMT+7). / When the error occurred.
      examples:
        - type: https://giapha.example.vn/problems/branch-scope-violation
          title: Ngoài phạm vi chi/ngành được giao
          status: 403
          detail: Bạn chỉ được sửa nhân khẩu thuộc chi Nhất.
          instance: /api/v1/persons/6f1b1d6e-2f8a-4d1c-9a91-9f2c1f0b7a01
          code: BRANCH_SCOPE_VIOLATION
          traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
...
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          description: |-
            Problem kèm chi tiết lỗi từng trường, cho FE tô đỏ đúng ô nhập.
          properties:
            errors:
              type: array
              description: Danh sách lỗi theo trường. / Per-field violations.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    description: |-
                      Đường dẫn trường theo JSON Pointer rút gọn, ví dụ `names[0].fullName`.
                  message:
                    type: string
                    description: Mô tả lỗi cho người dùng. / Human-readable message.
                  rejectedValue:
                    description: |-
                      Giá trị bị từ chối. **Bỏ trống với dữ liệu Tầng 3** (SĐT, email, địa chỉ) —
                      không dội giá trị nhạy cảm trở lại và cũng không ghi vào log.
...
    ReminderGenerationResult:
      type: object
      description: |-
        Kết quả một lượt chạy tay job sinh lịch nhắc giỗ.

        Đọc cặp `scannedEvents` / `createdJobs` cùng nhau: `createdJobs = 0` mà
        `scannedEvents > 0` nghĩa là **lịch nhắc đã có sẵn từ lượt trước** (chống trùng đang chạy
        đúng), còn `scannedEvents = 0` nghĩa là không có sự kiện lặp lại nào để quét.
      required: [referenceDate, scannedEvents, createdJobs]
      properties:
        referenceDate:
          type: string
          format: date
          description: Ngày được coi là "hôm nay" của lượt chạy (GMT+7). / Reference "today".
        lunarYears:
          type: array
          description: |-
            Các năm **âm lịch** đã quy đổi. Luôn gồm năm âm hiện tại và năm kế tiếp: giỗ tháng Chạp
            rơi vào tháng 1–2 **dương lịch của năm sau**, chỉ sinh cho năm âm hiện tại thì nó chỉ có
            lịch nhắc sau khi đã qua.
          items:
            type: integer
          examples: [[2026, 2027]]
        solarYears:
          type: array
          description: Các năm **dương lịch** đã quy đổi (cho sự kiện theo dương lịch).
          items:
            type: integer
          examples: [[2026, 2027]]
        scannedEvents:
          type: integer
          description: Số sự kiện lặp lại, chưa xóa mềm, đã quét. / Recurring events scanned.
        createdJobs:
          type: integer
          description: |-
            Số `reminder_job` **thực sự ghi mới**. Job đã tồn tại không được tính — đó là điều làm
            cho endpoint này idempotent.
        startedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          format: int64
          description: Thời gian chạy, mili giây. / Wall-clock duration.
    ReminderDispatchResult:
      type: object
      description: |-
        Kết quả một lượt đẩy lịch nhắc đã tới giờ lên RabbitMQ.
      required: [dispatchedJobs, dispatchedAt]
      properties:
        dispatchedJobs:
          type: integer
          description: |-
            Số `reminder_job` đã đẩy lên hàng đợi. **Không phải** số người đã nhận được: việc gửi
            thật diễn ra bất đồng bộ ở consumer, và luồng HTTP không bao giờ chờ nó.
            Job quá hạn hoặc thuộc sự kiện đã xóa mềm bị chuyển `CANCELLED` và không tính vào đây.
        dispatchedAt:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
    # WEB PUSH
    # -------------------------------------------------------------------------
    VapidPublicKey:
      type: object
      required: [publicKey]
      properties:
        publicKey:
          type: string
          description: |-
            Khóa công khai VAPID, **base64url không padding**, truyền thẳng vào
            `applicationServerKey` của `PushManager.subscribe`.
          examples: [BEl62iUYgUivxIkv69yViEuiBIa-Ib9-SkvMeAtA3LFgDzkrxZJjSgSnfckjBJuBkr3qBUYIHBQFLXYp5Nksh8U]
    PushSubscriptionCreateRequest:
      type: object
      title: Đăng ký Web Push / Web Push subscription
      description: |-
        Chính là `PushSubscription.toJSON()` của trình duyệt (chuẩn Web Push / RFC 8291),
        cộng vài trường phụ trợ để quản trị thiết bị.
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
          format: uri
          description: |-
            Endpoint do push service của trình duyệt cấp. **Đây là khóa định danh duy nhất**
            của một đăng ký — gửi lại cùng `endpoint` là cập nhật, không tạo bản mới.
        keys:
          type: object
          required: [p256dh, auth]
          description: Khóa mã hóa payload (RFC 8291). / Payload encryption keys.
          properties:
            p256dh:
              type: string
              description: Khóa công khai ECDH của client, base64url. / Client ECDH public key.
            auth:
              type: string
              description: Bí mật xác thực, base64url. / Auth secret.
        expirationTime:
          type: [integer, 'null']
          format: int64
          description: |-
            Thời điểm hết hạn do trình duyệt cấp (epoch milliseconds), thường `null`.
        userAgent:
          type: [string, 'null']
          description: |-
            User-Agent để người dùng nhận ra thiết bị trong màn hình cài đặt
            ("iPhone của Bố"). Không dùng cho mục đích khác.
        locale:
          type: [string, 'null']
          description: Ngôn ngữ mong muốn của thông báo đẩy tới thiết bị này (`vi`/`en`).
          examples: [vi]
    PushSubscriptionDto:
      type: object
      required: [id, endpoint, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: Id bản ghi — dùng cho `DELETE /push/subscriptions/{id}`.
        endpoint:
          type: string
          format: uri
          description: |-
            Endpoint đã đăng ký. **Không trả về `keys`** — khóa mã hóa chỉ nằm ở server,
            trả ra là rò rỉ vô ích.
        userAgent:
          type: [string, 'null']
        locale:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: [string, 'null']
          format: date-time
          description: Lần cuối gửi push thành công qua đăng ký này. / Last successful push.
        isCurrentDevice:
          type: boolean
          default: false
          description: |-
            Đăng ký này khớp `endpoint` mà client vừa gửi lên — giúp màn hình cài đặt đánh dấu
            "thiết bị này". / Marks the caller's own device.

    # -------------------------------------------------------------------------
    # GRAPHQL TRANSPORT
    # -------------------------------------------------------------------------
...